Fix #8: Ensure border covers the whole page.
I noticed for the two failing pages, an article on wikipedia.org and walmart.com, they both had a 'height: 100%' style on their body elements.

I saw two options to fixing this:
1. Keep the label appended to the body element and change its 'height' to 'unset' ('this.document.body.style.height = "unset";'), which seems risky, as it could break other pages.
2. Come up with a solution that doesn't involve modifying styles on existing elements on the page.

I opted for option 2, though this solution makes the border wrap around the viewport, rather than the body element. If anyone has strong feelings either way, please let me know.

I also made some tweaks to the styles for the label to make sure whitespace is even and the font style is consistent page to page, and added the category names to the console.logs for clarity.